+++ cfripper/rules/sns_topic_policy.py
@@ -17,6 +17,15 @@ class SNSTopicPolicyNotPrincipalRule(ResourceSpecificRule):
         AWS **strongly** recommends against using `NotPrincipal` in the same policy statement as `"Effect": "Allow"`.
         Doing so grants the permissions specified in the policy statement to all principals except the one named
         in the `NotPrincipal` element. By doing this, you might grant access to anonymous (unauthenticated) users.
+
+    Filters context:
+        | Parameter               | Type                             | Description                                                    |
+        |:-----------------------:|:--------------------------------:|:--------------------------------------------------------------:|
+        |`config`                 | str                              | `config` variable available inside the rule                    |
+        |`extras`                 | str                              | `extras` variable available inside the rule                    |
+        |`logical_id`             | str                              | ID used in Cloudformation to refer the resource being analysed |
+        |`resource`               | `SNSTopicPolicy`                 | Resource that is being addressed                               |
+        |`statement`              | `Statement`                      | Statement being checked found in the Resource                  |
     """
 
     GRANULARITY = RuleGranularity.RESOURCE
@@ -27,7 +36,18 @@ class SNSTopicPolicyNotPrincipalRule(ResourceSpecificRule):
         result = Result()
         for statement in resource.Properties.PolicyDocument._statement_as_list():
             if statement.NotPrincipal:
-                self.add_failure_to_result(result, self.REASON.format(logical_id), resource_ids={logical_id})
+                self.add_failure_to_result(
+                    result,
+                    self.REASON.format(logical_id),
+                    resource_ids={logical_id},
+                    context={
+                        "config": self._config,
+                        "extras": extras,
+                        "logical_id": logical_id,
+                        "resource": resource,
+                        "statement": statement,
+                    },
+                )
         return result