+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Implements `pluggy` https://github.com/pytest-dev/pluggy to enable dynamic rule loading.
 
+## [0.23.3] - 2021-02-11
+### Additions
+- All rules now support filter contexts!
+### Improvements
+- Update `WildcardResourceRule` to allow for certain resources to be excluded.
+
 ## [0.23.2] - 2021-02-04
 ### Bugfix
 - `GenericWildcardPrincipalRule` to ignore account IDs where full or partial wildcard is required in the Principal.